refactor to use hook to get te m members data
