+++ graylog2-web-interface/src/components/roles/RolesOverview/RolesOverview.tsx
@@ -15,19 +15,20 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 import * as React from 'react';
-import { useEffect, useState } from 'react';
+import { useEffect, useState, useCallback } from 'react';
 import styled from 'styled-components';
 import { PluginStore } from 'graylog-web-plugin/plugin';
 
 import AuthzRolesDomain from 'domainActions/roles/AuthzRolesDomain';
 import type { PaginatedRoles } from 'actions/roles/AuthzRolesActions';
 import { AuthzRolesActions } from 'stores/roles/AuthzRolesStore';
-import { Spinner, PaginatedList } from 'components/common';
+import { DataTable, Spinner, PaginatedList, EmptyResult } from 'components/common';
 import { Col, Row } from 'components/bootstrap';
 
-import RolesTable from './RolesTable';
+import RolesOverviewItem from './RolesOverviewItem';
 import RolesFilter from './RolesFilter';
 
+const TABLE_HEADERS = ['Name', 'Description', 'Users', 'Actions'];
 const DEFAULT_PAGINATION = {
   page: 1,
   perPage: 10,
@@ -56,6 +57,16 @@ const StyledPaginatedList = styled(PaginatedList)`
   }
 `;
 
+const _headerCellFormatter = (header) => {
+  // eslint-disable-next-line react/destructuring-assignment
+  switch (header.toLowerCase()) {
+    case 'actions':
+      return <th className="actions text-right">{header}</th>;
+    default:
+      return <th>{header}</th>;
+  }
+};
+
 const _loadRoles = (pagination, setLoading, setPaginatedRoles) => {
   setLoading(true);
 
@@ -67,14 +78,11 @@ const _loadRoles = (pagination, setLoading, setPaginatedRoles) => {
 
 const _updateListOnRoleDelete = (perPage, query, setPagination) => AuthzRolesActions.delete.completed.listen(() => setPagination({ page: DEFAULT_PAGINATION.page, perPage, query }));
 
-const _headerCellFormatter = (header) => {
-  // eslint-disable-next-line react/destructuring-assignment
-  switch (header.toLowerCase()) {
-    case 'actions':
-      return <th className="actions text-right">{header}</th>;
-    default:
-      return <th>{header}</th>;
-  }
+const getUseTeamMembersHook = () => {
+  const defaultHook = () => ({ leading: false, users: [] });
+  const teamsPlugin = PluginStore.exports('teams');
+
+  return teamsPlugin?.[0]?.useTeamMembersByRole || defaultHook;
 };
 
 const RolesOverview = () => {
@@ -83,19 +91,27 @@ const RolesOverview = () => {
   const [pagination, setPagination] = useState(DEFAULT_PAGINATION);
   const { list: roles } = paginatedRoles || {};
   const { page, perPage, query } = pagination;
-  const teamsPlugin = PluginStore.exports('teams');
-  const RolesTableWithTeamMembers = teamsPlugin?.[0]?.RolesTableWithTeamMembers;
+  const useTeamMembersByRole = getUseTeamMembersHook();
+  const rolesUsersByTeam = useTeamMembersByRole();
 
   useEffect(() => _loadRoles(pagination, setLoading, setPaginatedRoles), [pagination]);
   useEffect(() => _updateListOnRoleDelete(perPage, query, setPagination), [perPage, query]);
+
+  const _rolesOverviewItem = useCallback((role) => {
+    const { id } = role;
+    const roleUsers = paginatedRoles?.context.users[id];
+    const users = rolesUsersByTeam.users[id]
+      ? [...rolesUsersByTeam.users[id], ...roleUsers]
+      : paginatedRoles?.context?.users[id];
+
+    return <RolesOverviewItem role={role} users={users} />;
+  }, [rolesUsersByTeam, paginatedRoles?.context]);
 
   if (!paginatedRoles) {
     return <Spinner />;
   }
 
-  const searchFilter = <RolesFilter onSearch={(newQuery) => setPagination({ ...pagination, query: newQuery, page: 1 })} />;
-
-  const RolesTableComponent = RolesTableWithTeamMembers || RolesTable;
+  const searchFilter = <RolesFilter onSearch={(newQuery) => setPagination({ ...pagination, query: newQuery, page: DEFAULT_PAGINATION.page })} />;
 
   return (
     <Container>
@@ -111,12 +127,18 @@ const RolesOverview = () => {
           <StyledPaginatedList activePage={page}
                                totalItems={paginatedRoles.pagination.total}
                                onChange={(newPage, newPerPage) => setPagination({ ...pagination, page: newPage, perPage: newPerPage })}>
-            <RolesTableComponent roles={roles}
-                                 setPagination={setPagination}
-                                 searchFilter={searchFilter}
-                                 pagination={pagination}
-                                 headerCellFormatter={_headerCellFormatter}
-                                 paginatedRoles={paginatedRoles} />
+            <DataTable id="roles-overview"
+                       className="table-hover"
+                       rowClassName="no-bm"
+                       headers={TABLE_HEADERS}
+                       headerCellFormatter={_headerCellFormatter}
+                       sortByKey="name"
+                       rows={roles.toJS()}
+                       noDataText={<EmptyResult>No roles have been found.</EmptyResult>}
+                       customFilter={searchFilter}
+                       dataRowFormatter={_rolesOverviewItem}
+                       filterKeys={[]}
+                       filterLabel="Filter Roles" />
           </StyledPaginatedList>
         </Col>
       </Row>
